+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - split seq-repo load function into new asset
-- apply all activity filters to all extractors with activities
+- apply all activity filters to all extractors(except synopse) with activities
 - refactor filter functionality to filter module
 
 ### Deprecated
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - removed RDMO unused extractor
+- BREAKING: remove all activity filter rules from settings
 
 ### Fixed